@@ -1,5 +1,5 @@
 module golang.org/x/term
 
-go 1.11
+go 1.17
 
-require golang.org/x/sys v0.0.0-20201119102817-f84b799fce68
+require golang.org/x/sys v0.0.0-20210615035016-665e8c7367d1
